Changes needed to compile on macOS

+++ jlm/llvm/opt/RvsdgTreePrinter.cpp
@@ -106,12 +106,12 @@ RvsdgTreePrinter::AnnotateNumRvsdgNodes(
           numSubregionNodes += annotateRegion(*subregion);
         }
 
-        annotationMap.AddAnnotation(structuralNode, { label, numSubregionNodes });
+        annotationMap.AddAnnotation(structuralNode, { label, (uint64_t)numSubregionNodes });
       }
     }
 
     auto numNodes = region.nnodes();
-    annotationMap.AddAnnotation(&region, { label, numNodes });
+    annotationMap.AddAnnotation(&region, { label, (uint64_t)numNodes });
 
     return numNodes;
   };
@@ -134,12 +134,12 @@ RvsdgTreePrinter::AnnotateNumMemoryStateInputsOutputs(
     auto argumentRange = region.Arguments();
     auto numMemoryStateArguments =
         std::count_if(argumentRange.begin(), argumentRange.end(), IsMemoryStateOutput);
-    annotationMap.AddAnnotation(&region, { argumentLabel, numMemoryStateArguments });
+    annotationMap.AddAnnotation(&region, { argumentLabel, (uint64_t)numMemoryStateArguments });
 
     auto resultRange = region.Results();
     auto numMemoryStateResults =
         std::count_if(resultRange.begin(), resultRange.end(), IsMemoryStateInput);
-    annotationMap.AddAnnotation(&region, { resultLabel, numMemoryStateResults });
+    annotationMap.AddAnnotation(&region, { resultLabel, (uint64_t)numMemoryStateResults });
 
     for (auto & node : region.Nodes())
     {
@@ -154,7 +154,7 @@ RvsdgTreePrinter::AnnotateNumMemoryStateInputsOutputs(
             numMemoryStateInputs++;
           }
         }
-        annotationMap.AddAnnotation(structuralNode, { inputLabel, numMemoryStateInputs });
+        annotationMap.AddAnnotation(structuralNode, { inputLabel, (uint64_t)numMemoryStateInputs });
 
         size_t numMemoryStateOutputs = 0;
         for (size_t n = 0; n < structuralNode->noutputs(); n++)
@@ -165,7 +165,9 @@ RvsdgTreePrinter::AnnotateNumMemoryStateInputsOutputs(
             numMemoryStateOutputs++;
           }
         }
-        annotationMap.AddAnnotation(structuralNode, { outputLabel, numMemoryStateOutputs });
+        annotationMap.AddAnnotation(
+            structuralNode,
+            { outputLabel, (uint64_t)numMemoryStateOutputs });
 
         for (size_t n = 0; n < structuralNode->nsubregions(); n++)
         {
